Earth/Moon Complete
I exaggerated the bodies' sizes in order for them to be visible, and decoupled the masses from the calculated density (due to the large sizes). I'm doing this on the bus, so exact masses are unavailable. 

Next I will work on getting an orbit paramatizer and initial conditions setup, which will allow us to choose orbits at our leisure, as opposed to picking random velocities and radii, then hoping the Moon doesn't spiral into the Earth.