Working SI units
The code has been refactored to deal in SI units when giving units to roche.py, then converting the pixel values back into values which make sense. 

Currently the system is a mess, with incorrect values and not a comment in sight, but it at least gives a circular-ish orbit when I give it the Moon's orbital parameters!

(also added in the Sigurdson icon)